Add files via upload
A voice journaling app that transcribes your speech into text and also keeps the audio file, making it easy to journal on the go. The files are sorted by dates and subjects. The transcribed text can be edited, exported to text editor or compiled in sets to publish a book. The audio can be infused with background music to make into a podcast or playlist that you can listen to and share with friends. This project is for my junior frontend developer portfolio and is useful for everyone but particularly writers, poets, musicians, creatives and every lover of words.